experimental/chat_protocol/backend/chat_protocol_function/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azurefunctions-extensions-http-fastapi

openai
azure-identity
pydantic
pysignalr